Improve color quantization algorithm

+++ feature_camera_feed/src/main/java/com/stebitto/feature_camera_feed/data/ColorNameRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.stebitto.feature_camera_feed.data
 internal class ColorNameRepositoryImpl(
     private val colorNameRemoteSource: ColorNameRemoteSource
 ) : ColorNameRepository {
-    override suspend fun getColorName(red: Double, green: Double, blue: Double): Result<String> = runCatching {
+    override suspend fun getColorName(red: Float, green: Float, blue: Float): Result<String> = runCatching {
         return Result.success(
             colorNameRemoteSource
                 .getColorName(red.toInt(), green.toInt(), blue.toInt()).name.value
